ceph.spec.in: prefix conditionals with 0
Although "%if %..." works fine, prefixing with 0 is more robust.

For example, if we do not prefix and someone later decides to tack on
an "or" clause, they will be more likely to write "%if 0%... || 0%..."
(which is more likely to work everywhere) if we include the 0 prefix from
the beginning.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 6ead53225a0c7e5cb2b74478a0f6da7b569f056d)

Conflicts:
	ceph.spec.in
            take %{with lowmem_builder_override} (downstream-only) into account