packaging: add lsb-release dependency for SUSE and Debian
The lsb_release executable is being run in multiple places, not least in
src/common/util.cc, which calls it via shell in the collect_sys_info() code
path.

This patch addresses this issue on SUSE- and Debian-derivatives.

On RHEL/Fedora, lsb_release is in the redhat-lsb-core package. This used to
be a runtime dependency, but was dropped in
15600572265bed397fbd80bdd2b7d83a0e9bd918 and reinstating it is undesirable.

http://tracker.ceph.com/issues/14906 Fixes: #14906

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 011793733aae9c5b3ce4e36becdf996aa640cc5d)

Conflicts:
	ceph.spec.in
           trivial resolution