osdc/Objecter: fill out_handler/outbl for write ops
The new cmpext operation is bidirectional, in that comparison data is
carried in the OSD request, and mismatch data may be present in the OSD
response.
To ensure that cmpext mismatch data is processed by the context
completion callback in handle_osd_op_reply() for write operations,
e.g. rados_write_op_operate() dispatches, out_handler must be filled
accordingly.

Signed-off-by: David Disseldorp <ddiss@suse.de>