qa: tell install task to not install RPMs
For our DeepSea-based tests, we need the install task to add the repos,
but we do not need it to install all the packages listed in packages.yaml
(Ceph RPMs and their runtime dependencies), because DeepSea Stage 0 takes care
of installing those.

This has the positive side effect that install.py does not install the
ceph-test RPM, which is sometimes problematic to install due to a hard
versioned dependency in the spec file.

Signed-off-by: Nathan Cutler <ncutler@suse.com>