XXX libcephfs: add ceph_ll_walk_from()
Similar to ceph_ll_walk(), ceph_ll_walk_from() resolves a filesystem
path to an inode, but rather than starting from cwd or (for absolute
paths) root, ceph_ll_walk_from() takes a parent Inode parameter as the
starting point.

This functionality is useful for implementing behaviour similar to
unlinkat().

Signed-off-by: David Disseldorp <ddiss@suse.de>

TODO: What to do if the path is absolute - fail or ignore the parent